change getNote signature int to long

+++ app/src/main/java/com/tsci/notesapp/data/NoteDao.kt
@@ -16,7 +16,7 @@ interface NoteDao {
     suspend fun delete(note: Note)
 
     @Query("SELECT * FROM Note WHERE id = :id")
-    fun getNote(id: Int): Flow<Note>
+    fun getNote(id: Long): Flow<Note>
 
     @Query("SELECT * from Note ORDER BY note_date DESC")
     fun getAllNotes(): Flow<List<Note>>
